Add useNucleux hook (#2)
`useNucleux` provides access to all methods and state values from a
Nucleux store in a single hook.

This hook creates a reactive proxy to the store that automatically
updates when any atom in the store changes. It's designed for components
that need access to multiple atoms from the same store without having to
use multiple hooks.

Other changes:
- Add constraints to only accept stores that extend from a nucleux
store.
- Add documentation to all hooks.
- Add dependencies to `useStore` and `useValue`.
- Cleanup and test updates.